Fix so that it compiles with java 1.6
GNU Prolog for Java did not compile using Java 1.6 (mainly because of chang
es introduced in Java 1.4).
This patch fixes the code so that it compiles (with warnings) when using Ja
va 1.6. (It also breaks backwards compatibility because it uses generics: t
hat can be ignored if backwards compatibility is needed).

Signed-off-by: Daniel Thomas <drt24@cam.ac.uk>