Fix spelling mistakes in parser generators
TermParser.g had some obvious spelling mistakes which would have rendered it invalid and these are corrected.
In both TermParser.jj and TermParser.g SYMBOLIC was misspelt as SYMOLIC and so I have corrected that.

Add debug option to the generate-parser target in build.xml so that the local build.properties file can control 
that.

Signed-off-by: Daniel Thomas <drt24@cam.ac.uk>