Implement random/1 as defined in SWI-Prolog
The demo program mentalarithmetic requires a random(+IntExpr) evaluable predicate which is defined in
SWI-Prolog but not in ISO of GNU Prolog (which has it's own random predicates with different semantics).

This commit adds support for this predicate. However this can't be tested yet as other predicates are also
not yet implemented.

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>