Specify a sensible max arity
We can't possibly (with current machines) assume that we can create a
function with MAX_INT arguments. Instead make another gross assumption
and assume that we can use one byte per argument and use all possible
memory for that function.
(the motivation for doing this may relate to making it pass one of the
tests in the standard...)

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>