Fix some of the inria tests
Some of the inria tests were incorrect.

atom_codes assumed that 1000 was not a representable character code but
with UTF-8 this is no longer true.

catch-and-throw tried to throw something which would not be caught but
unfortunately the test harness has to catch it as part of the testing
framework and so the correct system_error cannot be thrown as intended.
the old test was moved to from_catc-and-throw. There was some sort of
parser bug when it was included but commented out in catch-and-throw.
(and we know that the parser needs fixing).

current_predicate had some tests added based on the ISO standard.

inria/delmemb was fixed as it was incorrect.

sub_atom was not specifying the values of After and although they are
not specified in the text in the standard they will be found by any
compliant implementation.

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>