Add tests of at_end_of_file
Import the tests mainly from the vanilla suite as the inria suite is
much nicer to work with. Now all the failing tests in the vanilla suite
are incorporated into the inria suite and so the vanilla suite can be
ignored during day-to-day development and only checked during
pre-release testing.

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>