Log errors during initialization and collect useful debug information
Previously there was an empty catch clause in Environment which meant
that when something went wrong it was hard to know what went wrong.

While fixing that it was realised that it was hard to tell where errors
generated using that came from and so more information was persisted in
case it was needed later when the initialization is run to report
errors.

In the process of doing this I used Pair<L,R> which has subsequently
been moved into the database package from demo.

Fixes were applied to Predicate_ensure_loaded to make it display errors
that occur during loading (to stderr) however there is a TODO: we need
to do a runInitialization() inside Predicate_ensure_loaded and to do
that we need to refactor Environment a bit first.

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>