Make runInitialization(Interpreter) reentrant
This allows us to properly support the use of ensure_loaded(file)
outside of the ISO directive context.

To do this instead of using a boolean flag to store whether we have done
it or not and then throwing an error if we are asked again we instead
have a list of the initializations and clear the list when we have
finished running them (so we don't run them again).

This also required some fairly dangerous changes to Interpreter as
ensure_loaded's runInitialization requires a new Goal to be set when one
already exists. To get around this we save the current state, execute
the new goal and when that finishes reset the state to the old state.
(this is dangerous as it might lead to subtle bugs).

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>