Improve testing
In arg add a minimal test case for the current failure.

In irinasuite switch to only using set_prolog_flag when we need to not
every time.

In at_end_of_stream use more parentheses to make it absolutely clear
what is happening.

In number_chars switch to a more obvious error in the input characters
so that parsers which are generous enough to parse '3 ' as 3 don't get
penalised.

In set_prolog_flag add more tests on the currently failing double_quotes
flag.

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>