Add regenerate-parser to build.xml
This should NOT be run at the moment as changes will have to be
reverted: it is there to help with the process of fixing the bug which
means that it can't be run namely that regenerating the parser results
in a non-terminating parser.
Also reorder tests to that inria goes first and load the
build.properties file so that we can have developer local build
properties.

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>