Fix TermParser.jj so that the code it produces compiles
So the code it produces is non terminating but this commit makes it
compile which is a start.

Mainly this refers to the change to using TermConstants.

ParseError.java was removed as it is never used and hasn't been for a
very long time.

Comment out an unused section from buildins.pro

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>