Fix bug in at_end_of_stream/1
TextInputPrologStream#getEndOfStreamState() was an unnecessary function
which was unfortunately completely wrong and always set endOfStreamState
to 'at' because termReader.ready() always returns false (I checked the
source) because it can never guarentee that calls to read will not
block. That is also not the question that termReader.ready() was trying
to answer.

The superclass method provides the correct answer.

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>