Eliminate redundant install() and unistall() methods
By using an ExecuteOnlyCode abstract class which has an empty method for
those methods.

This allowed the elimination of install and uninstall from all the
classes which implemented PrologCode except InterpretedByteCode.

I also removed the redundant javadoc which was copied and pasted around
it only needs to live in one place which is inside PrologCode.

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>