Fix DoubleQuotesTerm to use correct types
DoubleQuotesTerm must allow codesValue and charsValue to be AtomTerms
(as this is the case when we have an empty list) and so they must have
the type Term.

Check for null in ChangeableTerm and thow an IllegalArgumentException
there rather than a NullPointerException later as this is a slightly
better state of affairs.

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>