Fully support double_quotes
Tell the TermParser the Environment so that it can use it when it
constructs the DoubleQuotesTerm so that the DoubleQuotesTerm can pick
the correct value to have depending on the value of the double_quotes
flag at runtime.

This involves adding Environment to the constructors for
PrologTextLoaderState, TermReader and TermParser.

PrologStream.OpenOptions was given a constructor for the arguments which
are not intialized by default so that we could ensure that these were
being set correctly (as they were not being set correctly) some
redundancy was removed from TextInputPrologStream in the process.

One of the tests in set_prolog_flag was fixed (it was not quoting
numbers when using them as atoms).

HasEnvironment was created to encapsulate the getEnvironment() method
used in multiple classes.

DoubleQuotesTerm was fixed to correctly use the switch statement (I had
forgotten the breaks /o\)

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>