Add public key of the key signing key
The GPG key:
pub   4096R/A10C2B34 2010-07-26
      Key fingerprint = BEBF A393 37E8 2048 D3B3  28F4 72C9 4F71 A10C
2B34
uid                  GNU Prolog for Java (key signing key)
<info-gnuprologjava@gnu.org>
sub   4096R/C69491F1 2010-07-26

Is the key signing key for GNU Prolog for Java and will be used to sign
the GPG keys of developers who are allowed to do releases.

The public key is added to the sources and VCS as this makes it far
harder to destroy and embeds it in the history of the project and all
future commit hashes will depend on it.

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>