Add preliminary testing of extension to DCGs
Add preliminary testing of part 3 of the ISO standard and the provided
reference implementation (on the basis that if we can load it and pass
the tests then we must be doing fairly well at running Prolog and this
means that our users can know that we support part 3 without us having
to do much work).

Currently we don't pass these tests because they rely on the extensions
to part 1 of the ISO standard which have not yet been implemented.

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>