PrologException: handle evaluationError and improve syntaxError
Add an evaluationError(AtomTerm) method to be used by Evaluate so that
some of PrologException's internal representation can be hidden from
other classes.

Improve syntaxError(*) methods so that they give more details where
possible.

Catch a ParseException rather than an IOException so that we can do a
far better syntaxError.

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>