Refactor int return codes to enum RC and EXRC
PrologCode's SUCCESS, SUCCESS_LAST, FAIL, HALT static final int enum
pattern is refactored to an enum PrologCode.RC with the same fields.
This provides type safety and found a few horrible bits of code.

ExecutionState made similar use of int return codes and this was
refactored into an enum: ExecutionState.EXRC with the same fields as
before.

This is an APIBREAKING change. Next version MUST be a major version
increment (i.e. 0.3.0).

Signed-off-by: Daniel Thomas <drt24@srcf.ucam.org>