TermWriter changed to incorporate unicode writing.
Before the changes, TermWriter would print unicode in \x or \u form and
also would not recognize unicode characters and mark them as
needBeQuoted(unicode char). For this a new file UnicodeWriter.java has
also been created that has updated the definition of the identifiers in
GPJ. Doing so, makes the parser capable of handling around 19,000
unicode unicode characters.